move cpp/java to scenario folder
