+++ scenarios/cpp/windows/language-learning/README.md
@@ -1,0 +1,45 @@
+# C++ Console app for Windows
+
+This sample demonstrates various forms of speech recognition, intent recognition, conversation transcription and translation using the Speech SDK for C++ on Windows.
+
+> **Note:**
+> Conversation Transcription is in [Preview](https://aka.ms/cts/ctsoverview). The
+> ConversationTranscriber APIs are subjected to change. It requires
+> multi-microphone devices that can be paired with [Cognitive Services Speech
+> Device SDK](https://aka.ms/cts/getsdk).
+
+## Prerequisites
+
+* A subscription key for the Speech service. See [Try the speech service for free](https://docs.microsoft.com/azure/cognitive-services/speech-service/get-started).
+* A Windows PC with [Microsoft Visual Studio](https://www.visualstudio.com/) installed; some sample scenarios require a working microphone. See the [Speech SDK installation quickstart](https://learn.microsoft.com/azure/ai-services/speech-service/quickstarts/setup-platform?pivots=programming-language-cpp) for details on system requirements and setup.
+
+## Build the sample
+
+* **By building this sample you will download the Microsoft Cognitive Services Speech SDK. By downloading you acknowledge its license, see [Speech SDK license agreement](https://aka.ms/csspeech/license).**
+* [Download the sample code to your development PC.](/README.md#get-the-samples)
+* Start Microsoft Visual Studio and select **File** \> **Open** \> **Project/Solution**.
+* Navigate to the folder containing this sample, and select the solution file contained within it.
+* To tailor the sample to your configuration, use search and replace across the whole solution (for example, via **Edit** \> **Find and Replace** \> **Quick Replace**) to update the following strings:
+
+  * `YourSubscriptionKey`: replace with your subscription key.
+  * `https://YourServiceRegion.api.cognitive.microsoft.com`: replace with the endpoint for your Speech resource. You can find this endpoint in the Azure Portal under your Speech resource's "Keys and Endpoint" section.
+  * `YourEndpointId` (optional): replace with the endpoint ID of your customized model in [CRIS](https://cris.ai).
+* Set the active solution configuration and platform to the desired values under **Build** \> **Configuration Manager**:
+  * On a 64-bit Windows installation, choose `x64` as active solution platform.
+  * On a 32-bit Windows installation, choose `x86` as active solution platform.
+* Press Ctrl+Shift+B, or select **Build** \> **Build Solution**.
+
+> **Note**
+> If you are seeing red squigglies from IntelliSense for Speech SDK APIs,
+> right-click into your editor window and select **Rescan** > **Rescan Solution** to resolve.
+
+## Run the sample
+
+To debug the app and then run it, press F5 or use **Debug** \> **Start Debugging**. To run the app without debugging, press Ctrl+F5 or use **Debug** \> **Start Without Debugging**.
+
+The app displays a menu that you can navigate using your keyboard.
+Choose the scenarios that you're interested in.
+
+## References
+
+* [Speech SDK API reference for C++](https://aka.ms/csspeech/cppref)
