+++ scenarios/cpp/windows/language-learning/samples.sln
@@ -1,0 +1,30 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2011
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "samples", "samples\samples.vcxproj", "{6F0FEB3D-1411-4961-9BE0-CA0591077863}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Debug|x64.ActiveCfg = Debug|x64
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Debug|x64.Build.0 = Debug|x64
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Debug|x86.ActiveCfg = Debug|Win32
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Debug|x86.Build.0 = Debug|Win32
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Release|x64.ActiveCfg = Release|x64
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Release|x64.Build.0 = Release|x64
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Release|x86.ActiveCfg = Release|Win32
+		{6F0FEB3D-1411-4961-9BE0-CA0591077863}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9D62E0CE-8BF7-49DE-9B17-2FEC5A421F34}
+	EndGlobalSection
+EndGlobal
