+++ scenarios/cpp/windows/language-learning/samples/packages.config
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.CognitiveServices.Speech" version="1.45.0" targetFramework="native" />
+  <package id="Microsoft.CognitiveServices.Speech.Extension.MAS" version="1.45.0" targetFramework="native" />
+  <package id="nlohmann.json" version="3.10.4" targetFramework="native" />
+</packages>
