+++ scenarios/cpp/windows/language-learning/samples/stdafx.cpp
@@ -1,0 +1,8 @@
+// stdafx.cpp : source file that includes just the standard includes
+// CxxHelloWorld.pch will be the pre-compiled header
+// stdafx.obj will contain the pre-compiled type information
+
+#include "stdafx.h"
+
+// Reference any additional headers you need in stdafx.h
+// and not in this file
