+++ scenarios/cpp/windows/language-learning/samples/stdafx.h
@@ -1,0 +1,14 @@
+// stdafx.h : include file for standard system include files,
+// or project specific include files that are used frequently, but
+// are changed infrequently
+
+#pragma once
+
+#ifdef _WIN32
+#include "targetver.h"
+#include <tchar.h>
+#endif
+
+#include <stdio.h>
+#include <string.h>
+#include <iostream>
