+++ scenarios/java/jre/console/language-learning/.classpath
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+    <classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
+        <attributes>
+            <attribute name="maven.pomderived" value="true"/>
+        </attributes>
+    </classpathentry>
+    <classpathentry kind="src" output="target/classes" path="src">
+        <attributes>
+            <attribute name="optional" value="true"/>
+            <attribute name="maven.pomderived" value="true"/>
+        </attributes>
+    </classpathentry>
+    <classpathentry kind="con" path="org.eclipse.m2e.MAVEN2_CLASSPATH_CONTAINER">
+        <attributes>
+            <attribute name="maven.pomderived" value="true"/>
+        </attributes>
+    </classpathentry>
+    <classpathentry kind="src" path="target/generated-sources/annotations">
+        <attributes>
+            <attribute name="optional" value="true"/>
+            <attribute name="maven.pomderived" value="true"/>
+            <attribute name="ignore_optional_problems" value="true"/>
+            <attribute name="m2e-apt" value="true"/>
+        </attributes>
+    </classpathentry>
+    <classpathentry kind="src" output="target/test-classes" path="target/generated-test-sources/test-annotations">
+        <attributes>
+            <attribute name="optional" value="true"/>
+            <attribute name="maven.pomderived" value="true"/>
+            <attribute name="ignore_optional_problems" value="true"/>
+            <attribute name="m2e-apt" value="true"/>
+        </attributes>
+    </classpathentry>
+    <classpathentry kind="output" path="target/classes"/>
+</classpath>
